Final Project v1.1

+++ src/eshopper-backend-final/src/main/java/com/eshopper_backend_final/model/dto/ProductsDto.java
@@ -20,7 +20,7 @@ public class ProductsDto {
 
     private BigDecimal unitPrice;
 
-    private int unitsInStock;
+    private Long unitsInStock;
 
     private String picture;
 }
